requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.26.2
pyyaml==6.0.1
httpx==0.25.2
